+++ StortSpel/StortSpel/Material.h
@@ -9,7 +9,7 @@ struct MATERIAL_CONST_BUFFER
 	float UVScale			= 1.0f;
 	float roughness			= 0.5f;
 	float metallic			= 0.f;
-	int textured			= 0.f;
+	int textured			= 0;
 	float emissiveStrength	= 0.f; // 0 to 100
 	XMFLOAT3 pad;
 };
@@ -39,7 +39,7 @@ public:
 	void setRoughness(float roughness);
 	void setMetallic(float metallic);
 	void setTextured(int textured);
-	void setEmissiveStrength(int emissiveStrength);
+	void setEmissiveStrength(float emissiveStrength);
 
 	unsigned int long getMaterialId();
 	MATERIAL_CONST_BUFFER getMaterialParameters();
